# docs/swagger.yaml
basePath: /
definitions:
...
  genproto.CreateTransactionRequest:
    properties:
      account_id:
        type: number
      amount:
        type: string
      category_id:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Response:
    properties:
      message:
        type: string
    type: object
...
  genproto.TransactionDeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  genproto.TransactionResponse:
    properties:
      account_id:
        type: number
      amount:
        type: string
      category_id:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.TransactionsResponse:
    properties:
      transactions:
        items:
          $ref: '#/definitions/genproto.TransactionResponse'
        type: array
    type: object
...
  genproto.UpdateTransactionRequest:
    properties:
      account_id:
        type: number
      amount:
        type: string
      category_id:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Auth service API documentation
  title: api gat way
  version: "1.0"
paths:
...
  /transaction/create:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: Create transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction created successfully
          schema:
            $ref: '#/definitions/genproto.Response'
        "500":
          description: Error while creating transaction
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Transaction
      tags:
      - Transaction
  /transaction/delete/{transaction_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing transaction by its ID
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            $ref: '#/definitions/genproto.TransactionDeleteResponse'
        "400":
          description: 'Missing required query parameter: transaction_id'
          schema:
            type: string
        "500":
          description: Error while deleting transaction
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Transaction
      tags:
      - Transaction
  /transaction/get/{transaction_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific transaction's details by its ID
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction details
          schema:
            $ref: '#/definitions/genproto.TransactionResponse'
        "400":
          description: 'Missing required query parameter: transaction_id'
          schema:
            type: string
        "500":
          description: Error while fetching transaction
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Transaction by ID
      tags:
      - Transaction
  /transaction/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all transactions
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            $ref: '#/definitions/genproto.TransactionsResponse'
        "500":
          description: Error while retrieving transactions
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Transactions
      tags:
      - Transaction
  /transaction/update:
    put:
      consumes:
      - application/json
      description: Update the details of an existing transaction
      parameters:
      - description: Update transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction updated successfully
          schema:
            $ref: '#/definitions/genproto.Response'
        "400":
          description: Invalid request or missing transaction_id
          schema:
            type: string
        "500":
          description: Error while updating transaction
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Transaction
      tags:
      - Transaction
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"